# EXPORT APIDOG.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My Project",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/miembro/miembros": {
      "post": {
        "summary": "Guardar un Miembro",
        "x-apidog-folder": "POO2023/Miembros",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "POO2023/Miembros"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "tipo": "alumno",
                "nombre": "Roberto",
                "correo": "asdasd@gmail.com",
                "numeroTelefonico": "381616064",
                "libretaUniversitaria": "APU4073"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/412546/apis/api-4673060-run",
        "security": []
      },
      "get": {
        "summary": "Buscar Por Correo",
        "x-apidog-folder": "POO2023/Miembros",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "POO2023/Miembros"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "correo": {
                    "type": "string",
                    "example": "enzo.meneghini@hotmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Miembro": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "numeroMiembro": {
                          "type": "integer"
                        },
                        "nombre": {
                          "type": "string"
                        },
                        "correo": {
                          "type": "string"
                        },
                        "numeroTelefonico": {
                          "type": "string"
                        },
                        "fechaBloqueo": {
                          "type": "string"
                        },
                        "legajo": {
                          "type": "string"
                        },
                        "alumno": {
                          "type": "boolean"
                        },
                        "docente": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "numeroMiembro",
                        "nombre",
                        "correo",
                        "numeroTelefonico",
                        "fechaBloqueo",
                        "legajo",
                        "alumno",
                        "docente"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "numeroMiembro",
                        "nombre",
                        "correo",
                        "numeroTelefonico",
                        "fechaBloqueo",
                        "legajo",
                        "alumno",
                        "docente"
                      ]
                    }
                  },
                  "required": [
                    "Miembro"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "Miembro"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/412546/apis/api-4685972-run",
        "security": []
      }
    },
...
    "/api/v1/miembro/miembros/": {
      "put": {
        "summary": "Modificar todo un miembro",
        "x-apidog-folder": "POO2023/Miembros",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "POO2023/Miembros"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "tipo": "docente",
                "id": 1,
                "numeroMiembro": 15,
                "nombre": "Enzo",
                "correo": "enzo.meneghini@hotmail.com",
                "numeroTelefonico": "3885485126",
                "fechaBloqueo": "11/11/2023 - 09:00",
                "legajo": "FI0025"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Mensaje": {
                      "type": "string"
                    },
                    "Objeto": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "numeroMiembro": {
                          "type": "integer"
                        },
                        "nombre": {
                          "type": "string"
                        },
                        "correo": {
                          "type": "string"
                        },
                        "numeroTelefonico": {
                          "type": "string"
                        },
                        "fechaBloqueo": {
                          "type": "string"
                        },
                        "legajo": {
                          "type": "string"
                        },
                        "docente": {
                          "type": "boolean"
                        },
                        "alumno": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "numeroMiembro",
                        "nombre",
                        "correo",
                        "numeroTelefonico",
                        "fechaBloqueo",
                        "legajo",
                        "docente",
                        "alumno"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "numeroMiembro",
                        "nombre",
                        "correo",
                        "numeroTelefonico",
                        "fechaBloqueo",
                        "legajo",
                        "docente",
                        "alumno"
                      ]
                    }
                  },
                  "required": [
                    "Mensaje",
                    "Objeto"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "Mensaje",
                    "Objeto"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/412546/apis/api-4686372-run",
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": []
}